Reorganize default ufunc dispatcher definitions
